backend/requirements.txt
fastapi>=0.135,<1
# python-multipart нужен FastAPI для OAuth2PasswordRequestForm в /auth/login.
# В старых версиях FastAPI тянулся транзитивно, с 0.95+ стал опциональным —
# без него импорт app падает с RuntimeError ещё на старте (эндпоинт с формой
# проверяет наличие multipart в момент регистрации роута).
# Минимальная версия 0.0.18 — там пофикшен CVE-2024-53981 (DoS через
# длинный boundary в multipart-заголовке).
python-multipart>=0.0.18,<1
uvicorn[standard]>=0.42,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.31,<1
pydantic[email]>=2.12,<3
python-dotenv>=1.2,<2
bcrypt>=5,<6
email-validator>=2.3,<3
httpx>=0.28,<1
sentry-sdk[fastapi]>=2.0,<3
python-jose[cryptography]>=3.3,<4
alembic>=1.14,<2